Deny admin permission to eliminate or change role of other users

+++ api/src/controllers/user.js
@@ -191,37 +191,49 @@ module.exports = {
   },
   deleteUser: async (req, res) => {
     try {
+      const decodedToken = await verifyToken(req, res)
+      const userWillingToDelete = await User.findOne({ where: { id: decodedToken.id  } })
+      console.log("userWillingToDelete", userWillingToDelete.dataValues.email) //
       const { email } = req.params
-      const user = await User.findOne({ where: { email } })
-      const removedUser = await User.destroy({
-        where: { email },
-      })
-      if (removedUser) {
-        await Review.destroy({ where: { userId: user.id } })
-        return res.send({ msg: `User ${email} removed` })
+      if(userWillingToDelete.dataValues.email === "admin@gmail.com"){
+        const user = await User.findOne({ where: { email } })
+        const removedUser = await User.destroy({
+          where: { email },
+        })
+        if (removedUser) {
+          await Review.destroy({ where: { userId: user.id } })
+          return res.send({ msg: `User ${email} removed` })
+        }
+        return res.status(400).send({
+          error: `User ${email} doesnt exist`,
+        })
       }
-      return res.status(400).send({
-        error: `User ${email} doesnt exist`,
-      })
+      return res.send({ msg: `User ${userWillingToDelete.dataValues.email} has no permitions to delete other users` })
+
     } catch (error) {
       sendError(res, error)
     }
   },
   changeUsersRole: async (req, res) => {
     try {
+      const decodedToken = await verifyToken(req, res)
+      const userWillingToChangeRole = await User.findOne({ where: { id: decodedToken.id  } })
       const { email, adminState } = req.body
-      const foundUser = await User.findOne({
-        where: { email },
-      })
-      if (foundUser) {
-        foundUser.isAdmin = adminState
-        foundUser.save()
-        return res.send(
-          `${email} was changed to ${adminState ? "Admin" : "User"}`
-        )
-      } else {
-        return res.send("The email passed was not found")
+      if(userWillingToChangeRole.dataValues.email === "admin@gmail.com"){
+        const foundUser = await User.findOne({
+          where: { email },
+        })
+        if (foundUser) {
+          foundUser.isAdmin = adminState
+          foundUser.save()
+          return res.send(
+            `${email} was changed to ${adminState ? "Admin" : "User"}`
+          )
+        } else {
+          return res.send("The email passed was not found")
+        }
       }
+      return res.send({ msg: `User ${userWillingToChangeRole.dataValues.email} has no permitions to change other users role` })
     } catch (error) {
       return sendError(res, error)
     }
